Update game rules and fix invalid input bug
Previously the game asked how many games would be played, and the player
with the most wins after that number of games is played is the winner.
Now the game asks what the winning score will be, and the first player
to reach that number is the winner.

A bug in player choice caused issues with input validation. Previously,
if invalid input was enteered and then valid input was entered into the
error prompt, the program owuld not register the valid input. The new
input was not being checked in the proper format, so the function was
edited to capitalize the first letter and lower the rest.