Add selection logic and more styling to gamepieces
Wrote JS code to add click event listener to each gamepiece image, so
that when one is hovered over it inflates, and when selected it is
highlighted. Clicking again remnoves the highlight, and clicking on
another piece switches the highlight to that one.

Also added a media query to shrink images when screen size is too small.